Session info

Wrapper for session_info.

session_info(collapse = FALSE, title = "details", add_line = TRUE, ...)

Source

quarto example

Arguments

collapse

Wrap the session info output in a collapsible HTML element.

title

Title of the collapsed element.

add_line

Add a divider line after the session info.

...

Arguments passed on to sessioninfo::session_info

pkgs

Which packages to show. It may be:

NULL or "loaded": show all loaded packages,

"attached": show all attached packages,

"installed": show all installed packages,

a character vector of package names. Their (hard) dependencies are also shown by default, see the dependencies argument.

include_base

Include base packages in summary? By default this is false since base packages should always match the R version.

info

What information to show, it can be "auto" to choose automatically, "all" to show everything, or a character vector with elements from:

"platform": show platform information via platform_info(),

"packages": show package information via package_info(),

"python": show Python configuration via python_info(),

"external": show information about external software, via external_info().

dependencies

Whether to include the (recursive) dependencies as well. See the dependencies argument of utils::install.packages().

to_file

Whether to print the session information to a file. If TRUE the name of the file will be session-info.txt, but to_file may also be a string to specify the file name.

Value

Null.

Examples

session_info()
#> ─ Session info ───────────────────────────────────────────────────────────────
#>  setting  value
#>  version  R Under development (unstable) (2023-11-22 r85609)
#>  os       Ubuntu 22.04.3 LTS
#>  system   x86_64, linux-gnu
#>  ui       X11
#>  language en
#>  collate  C
#>  ctype    en_US.UTF-8
#>  tz       UTC
#>  date     2023-12-10
#>  pandoc   3.1.1 @ /usr/local/bin/ (via rmarkdown)
#> 
#> ─ Packages ───────────────────────────────────────────────────────────────────
#>  package       * version    date (UTC) lib source
#>  Biobase         2.63.0     2023-10-24 [1] Bioconductor
#>  BiocBaseUtils   1.5.0      2023-10-24 [1] Bioconductor
#>  BiocCheck       1.39.0     2023-10-24 [1] Bioconductor
#>  BiocFileCache   2.11.1     2023-10-26 [1] Bioconductor
#>  BiocGenerics    0.49.1     2023-11-01 [1] Bioconductor
#>  BiocManager     1.30.22    2023-08-08 [1] CRAN (R 4.4.0)
#>  biocViews       1.71.1     2023-10-25 [1] Bioconductor
#>  bit             4.0.5      2022-11-15 [1] CRAN (R 4.4.0)
#>  bit64           4.0.5      2020-08-30 [1] CRAN (R 4.4.0)
#>  bitops          1.0-7      2021-04-24 [1] CRAN (R 4.4.0)
#>  blob            1.2.4      2023-03-17 [1] CRAN (R 4.4.0)
#>  cachem          1.0.8      2023-05-01 [2] CRAN (R 4.4.0)
#>  cli             3.6.1      2023-03-23 [2] CRAN (R 4.4.0)
#>  codetools       0.2-19     2023-02-01 [3] CRAN (R 4.4.0)
#>  colorspace      2.1-0      2023-01-23 [1] CRAN (R 4.4.0)
#>  curl            5.2.0      2023-12-08 [2] CRAN (R 4.4.0)
#>  data.table      1.14.10    2023-12-08 [1] CRAN (R 4.4.0)
#>  DBI             1.1.3      2022-06-18 [1] CRAN (R 4.4.0)
#>  dbplyr          2.4.0      2023-10-26 [1] CRAN (R 4.4.0)
#>  desc            1.4.3      2023-12-10 [1] CRAN (R 4.4.0)
#>  devoptera     * 0.99.1     2023-12-10 [1] Bioconductor
#>  devtools        2.4.5      2022-10-11 [2] CRAN (R 4.4.0)
#>  digest          0.6.33     2023-07-07 [2] CRAN (R 4.4.0)
#>  downlit         0.4.3      2023-06-29 [2] CRAN (R 4.4.0)
#>  dplyr           1.1.4      2023-11-17 [1] CRAN (R 4.4.0)
#>  ellipsis        0.3.2      2021-04-29 [2] CRAN (R 4.4.0)
#>  evaluate        0.23       2023-11-01 [2] CRAN (R 4.4.0)
#>  fansi           1.0.6      2023-12-08 [2] CRAN (R 4.4.0)
#>  farver          2.1.1      2022-07-06 [1] CRAN (R 4.4.0)
#>  fastmap         1.1.1      2023-02-24 [2] CRAN (R 4.4.0)
#>  filelock        1.0.2      2018-10-05 [1] CRAN (R 4.4.0)
#>  fs              1.6.3      2023-07-20 [2] CRAN (R 4.4.0)
#>  generics        0.1.3      2022-07-05 [1] CRAN (R 4.4.0)
#>  ggfun           0.1.3      2023-09-15 [1] CRAN (R 4.4.0)
#>  ggimage         0.3.3      2023-06-19 [1] CRAN (R 4.4.0)
#>  ggplot2         3.4.4      2023-10-12 [1] CRAN (R 4.4.0)
#>  ggplotify       0.1.2      2023-08-09 [1] CRAN (R 4.4.0)
#>  glue            1.6.2      2022-02-24 [2] CRAN (R 4.4.0)
#>  graph           1.81.0     2023-10-24 [1] Bioconductor
#>  gridGraphics    0.5-1      2020-12-13 [1] CRAN (R 4.4.0)
#>  gtable          0.3.4      2023-08-21 [1] CRAN (R 4.4.0)
#>  here            1.0.1      2020-12-13 [1] CRAN (R 4.4.0)
#>  hexbin          1.28.3     2023-03-21 [1] CRAN (R 4.4.0)
#>  hexSticker      0.4.9      2020-12-05 [1] CRAN (R 4.4.0)
#>  htmltools       0.5.7      2023-11-03 [2] CRAN (R 4.4.0)
#>  htmlwidgets     1.6.4      2023-12-06 [2] CRAN (R 4.4.0)
#>  httpuv          1.6.13     2023-12-06 [2] CRAN (R 4.4.0)
#>  httr            1.4.7      2023-08-15 [2] CRAN (R 4.4.0)
#>  jsonlite        1.8.8      2023-12-04 [2] CRAN (R 4.4.0)
#>  knitr           1.45       2023-10-30 [2] CRAN (R 4.4.0)
#>  labeling        0.4.3      2023-08-29 [1] CRAN (R 4.4.0)
#>  later           1.3.2      2023-12-06 [2] CRAN (R 4.4.0)
#>  lattice         0.22-5     2023-10-24 [3] CRAN (R 4.4.0)
#>  lifecycle       1.0.4      2023-11-07 [2] CRAN (R 4.4.0)
#>  magick          2.8.1      2023-10-22 [1] CRAN (R 4.4.0)
#>  magrittr        2.0.3      2022-03-30 [2] CRAN (R 4.4.0)
#>  memoise         2.0.1      2021-11-26 [2] CRAN (R 4.4.0)
#>  mime            0.12       2021-09-28 [2] CRAN (R 4.4.0)
#>  miniUI          0.1.1.1    2018-05-18 [2] CRAN (R 4.4.0)
#>  munsell         0.5.0      2018-06-12 [1] CRAN (R 4.4.0)
#>  pillar          1.9.0      2023-03-22 [2] CRAN (R 4.4.0)
#>  pkgbuild        1.4.3      2023-12-10 [1] CRAN (R 4.4.0)
#>  pkgconfig       2.0.3      2019-09-22 [2] CRAN (R 4.4.0)
#>  pkgdown         2.0.7      2022-12-14 [1] CRAN (R 4.4.0)
#>  pkgload         1.3.3      2023-09-22 [2] CRAN (R 4.4.0)
#>  profvis         0.3.8      2023-05-02 [2] CRAN (R 4.4.0)
#>  promises        1.2.1      2023-08-10 [2] CRAN (R 4.4.0)
#>  purrr           1.0.2      2023-08-10 [2] CRAN (R 4.4.0)
#>  R.methodsS3     1.8.2      2022-06-13 [1] CRAN (R 4.4.0)
#>  R.oo            1.25.0     2022-06-12 [1] CRAN (R 4.4.0)
#>  R.utils         2.12.3     2023-11-18 [1] CRAN (R 4.4.0)
#>  R6              2.5.1      2021-08-19 [2] CRAN (R 4.4.0)
#>  ragg            1.2.6      2023-10-10 [2] CRAN (R 4.4.0)
#>  RBGL            1.79.0     2023-10-24 [1] Bioconductor
#>  Rcpp            1.0.11     2023-07-06 [2] CRAN (R 4.4.0)
#>  RCurl           1.98-1.13  2023-11-02 [1] CRAN (R 4.4.0)
#>  remotes         2.4.2.9000 2023-12-10 [1] Github (r-lib/remotes@6fe8836)
#>  rlang           1.1.2      2023-11-04 [2] CRAN (R 4.4.0)
#>  rmarkdown       2.25       2023-09-18 [1] CRAN (R 4.4.0)
#>  rprojroot       2.0.4      2023-11-05 [2] CRAN (R 4.4.0)
#>  RSQLite         2.3.4      2023-12-08 [1] CRAN (R 4.4.0)
#>  rstudioapi      0.15.0     2023-07-07 [2] CRAN (R 4.4.0)
#>  RUnit           0.4.32     2018-05-18 [1] CRAN (R 4.4.0)
#>  scales          1.3.0      2023-11-28 [1] CRAN (R 4.4.0)
#>  sessioninfo     1.2.2      2021-12-06 [2] CRAN (R 4.4.0)
#>  shiny           1.8.0      2023-11-17 [2] CRAN (R 4.4.0)
#>  showtext        0.9-6      2023-05-03 [1] CRAN (R 4.4.0)
#>  showtextdb      3.0        2020-06-04 [1] CRAN (R 4.4.0)
#>  stringdist      0.9.12     2023-11-28 [1] CRAN (R 4.4.0)
#>  stringi         1.8.2      2023-11-23 [2] CRAN (R 4.4.0)
#>  stringr         1.5.1      2023-11-14 [2] CRAN (R 4.4.0)
#>  sysfonts        0.8.8      2022-03-13 [1] CRAN (R 4.4.0)
#>  systemfonts     1.0.5      2023-10-09 [2] CRAN (R 4.4.0)
#>  textshaping     0.3.7      2023-10-09 [2] CRAN (R 4.4.0)
#>  tibble          3.2.1      2023-03-20 [2] CRAN (R 4.4.0)
#>  tidyselect      1.2.0      2022-10-10 [1] CRAN (R 4.4.0)
#>  urlchecker      1.0.1      2021-11-30 [2] CRAN (R 4.4.0)
#>  usethis         2.2.2      2023-07-06 [2] CRAN (R 4.4.0)
#>  utf8            1.2.4      2023-10-22 [2] CRAN (R 4.4.0)
#>  vctrs           0.6.5      2023-12-01 [2] CRAN (R 4.4.0)
#>  whisker         0.4.1      2022-12-05 [2] CRAN (R 4.4.0)
#>  withr           2.5.2      2023-10-30 [2] CRAN (R 4.4.0)
#>  xfun            0.41       2023-11-01 [2] CRAN (R 4.4.0)
#>  XML             3.99-0.16  2023-11-29 [1] CRAN (R 4.4.0)
#>  xml2            1.3.6      2023-12-04 [2] CRAN (R 4.4.0)
#>  xtable          1.8-4      2019-04-21 [2] CRAN (R 4.4.0)
#>  yaml            2.3.7      2023-01-23 [2] CRAN (R 4.4.0)
#>  yulab.utils     0.1.0      2023-09-20 [1] CRAN (R 4.4.0)
#> 
#>  [1] /__w/_temp/Library
#>  [2] /usr/local/lib/R/site-library
#>  [3] /usr/local/lib/R/library
#> 
#> ──────────────────────────────────────────────────────────────────────────────